+++ packages/core/src/web/helpers/convertToImage/types.ts
@@ -1,0 +1,13 @@
+import type { IBatchCommand } from '@core/interfaces/IHistory';
+
+export type ConvertSvgToImageParams = {
+  isToSelect?: boolean;
+  parentCmd?: IBatchCommand;
+  svgElement: SVGGElement;
+};
+export type ConvertToImageResult = undefined | { imageElements: SVGImageElement[]; svgElements: SVGGElement[] };
+
+export type BBox = { height: number; width: number; x: number; y: number };
+export type CreateImageParams = Record<'angle' | 'height' | 'width' | 'x' | 'y', number> &
+  Record<'href' | 'transform', string>;
+export type MainConverterFunc = (params: ConvertSvgToImageParams) => Promise<ConvertToImageResult>;
